Fix Typo in admin.php
When using a custom style and not creating '../resources/css/custom_chromakeying.css' the code tries to copy the default file, which results in an error because the path has duplicate .css and the file cannot be found: 

<br />
<b>Warning</b>:  copy(../resources/css/modern_chromakeying.css.css): failed to open stream: No such file or directory in <b>/var/www/html/api/admin.php</b> on line <b>117</b><br />
"success"

This results in the admin ui staying in the "saving" state indefinitely. Also it still returns "success" which might need to be changed.