core: fix preview from device cam at countdown
Inside api.startVideo we've only checked for the "view" mode if preview
from gphoto was enabled, else we always only used the "preview" mode.
Because of this a "view" of the device cam was not visible at the countdown.

The bug was introduced in commit 97fbf1c7.

We should consider renaming some vars in future for an easier understanding
while looking at the code.

Change-Id: I5791b870046dc5be4a7dcf0bd1aeda6434137f33