task: bump node version to v18 LTS and drop yarn
Using yarn as an alternative package manager to npm
that is already shipped with nodejs makes the installation
process unnecessary and complex.

Back in the days when the package-lock.json was not
respected by npm, yarn did a good job of installing
the same packages as defined in the lock file. Since
npm nowadays respects the lock file, there is no real
reason anymore for promoting/using yarn.

All scripts can be run as before the only difference
is that instead of yarn <scriptname> now
npm run <scriptname> needs to be called.

We are also bumping the necessary node version to
the current LTS 18.17.x. Advanced users with
Multiple node version requirements a .nvmrc file is
now shipped to help users select the correct node
version using nvm.

see: https://github.com/nvm-sh/nvm