task: introduce composer as php dependency manager
Dependencies in the php universe are managed via
composer, which also provides autoloading of classes
and several other benefits like platform checks etc.
See: https://getcomposer.org/

To avoid current installations breaking and introducing
a new package that needs to be installed we are starting
shipping the binary with the package for now until we
move over to a more sophisticated setup.

Composer is currently only used to install additional tooling
for PHP CGL fixing and checking and also PHP linting. This
has no direct effect on users and is only relevant for
development purposes right now.

We are introducing 2 packages here for the start:

1. PHP-CS-Fixer: https://github.com/PHP-CS-Fixer/PHP-CS-Fixer
    We are using PHP-CS-Fixer to ensure PHP CGL and enforce the
    PHP-FIG (PSR-12) standard recommendation for code formatting.
    See: https://www.php-fig.org/psr/

2. PHPLint: https://github.com/overtrue/phplint
    We are using PHPLint to simply PHP files are valid and do not
    contain any syntax errors.

The installation scripts for the tools are merged into the current
build step `npm run build` so dependencies are updated and
handled as needed without the interaction of a user required.

During the normal build process, the Coding Guidelines and
Linting checks are not executed but can be called manually.

1. Through NPM:
   - `npm run composer:install`
   - `npm run composer:update`
   - `npm run composer:cgl`
   - `npm run composer:lint`
   - `npm run composer:self-update`

2. Direct Binary access:
   - `bin/composer install`
   - `bin/composer update`
   - `bin/composer cgl`
   - `bin/composer lint`
   - `bin/composer self-update`

4. Global Composer installation (https://getcomposer.org/)
   - `composer install`
   - `composer update`
   - `composer cgl`
   - `composer lint`
   - `composer self-update`