task: introduce autoloading
With this patch, we are starting to introduce autoloading to the PHP
part of the application. With the introduction of composer, we can
make use of PSR-4 (https://www.php-fig.org/psr/psr-4/). This enables
us to avoid manually including PHP class files. Instead, we can import
them via namespaced call and the autoloader will provide the correct
resources for us.

All lib classes are now PSR-4 compatible and the new namespace
`Photobooth` was introduced. A new `lib\boot.php` file was introduced
that provides all the globally defined constants and loads the config.
PHP-Scripts that include the boot file will now have access to
autoloading, there is no need anymore for manual includes if the class/
package supports PSR-4.

The package `phpmailer/phpmailer` is now maintained via composer
and the git submodule was removed.