task: replace legacy phpqrcode with endroid/qr-code
The current phpqrcode does not support PSR-4 autoloading
and is not namespaced. We are switching to endroid/qr-code
to ease the generation of creating qrcodes. The dependency
is now managed via composer.

The current phpqrcode does not support PSR-4 autoloading
and is not namespaced. We are switching to endroid/qr-code
to ease the generation of creating QR codes.

We are dropping the config error correction level from the
configuration, this is too advanced for general users and
without deep knowledge of QR codes the levels are chosen
by looks and not by its usefulness.

The error correction level describes how damaged a
QR code can be before it is not readable anymore. Since
we are using screens/printouts, we do not expect a
high damage rate like on packaging materials.

The error correction is now by default medium (~15%).