task: streamline path handling (#324)
This patch aims to streamline the path handling.
The variable $fileRoot was passed through includes
multiple times and includes relying on it to be correct.

We are now removing this variable and all
manually set/assumed paths. With the
introduction of the PathUtility, all paths can be
correctly and reliably calculated. It auto-detects
subfolder installations and adjusts the paths
accordingly.