task: replace javascript language handling through php
The javascript solution for translation handling introduced an
unavoidable layout shift directly after the load when the labels
were replaced. To avoid this shift the translations are now handled
through PHP and are shipped directly to the browser without and
replacement necessary.

A new API endpoint `/api/translations.php` is introduced that is
used for the javascript side, where translations are fetched and
displayed on demand.