bugfix: rework imaging flow
Since the overflow of the aperture is not hidden
the effect leads to a wrong-scaled interface for
a short period of time. This patch resolves the
issue by ensuring the animation does not overflow
the viewport.

The countdown, cheese message and the shutter are
now async and functions that wait to be completed
before starting another action.

In addition, we are moving the complete handling
into the javascript, so the elements do not need
to be placed in the templates anymore.

run: in the browser command line:
photoBooth.countdown.start(10);
photoBooth.cheese.start();
photoBooth.shutter.start();