task: drop browsersync and fetch polyfill
In order to switch do a bundler at some point
we are starting to clean up dependencies.

The browser sync only runs under certain
conditions and is no replacement for a real
watcher. It has a lot of dependencies that
are affected from security issues.

Just run `npm run build` for now.

The fetch polyfill is not nessesary anymore,
the fetch api is widely supported.