task: drop browsersync and fetch polyfill (#346)
In order to switch to a bundler at some point
we are starting to clean up dependencies.

The browser sync only runs under certain
conditions and is no replacement for a real
watcher. It has a lot of dependencies that
are affected by security issues.

Just run `npm run build` for now.

The fetch polyfill is not necessary anymore,
the fetch API is widely supported.