task: windows improvements
The command `rm` is not supported on windows, we are replacing the
npm commands with the rimraf to support windows and unix.

Windows does not understand the direct execution of php scripts,
we are adding the bin `php` so windows can also handle the composer
scripts.

Adjusted the `build:head` command to also run on windows.