task: cleanup gallery
the gallery is used in multiple places:
- Startpage
- Standalone gallery
- Chroma
- Slideshow, sort of

Only the startpage allowed additional actions
directly on the gallery, like taking photos etc.
We are facing the problem here that we have
limited space available so we are dropping the
the actions on this view and it has to be closed
before the user can continue to take pictures.
This frees up more space to view the images
in the gallery and does not mix up functions
anymore.

The gallery should now load more smooth
and is already loaded on views like standalone
gallery and slideshow, this avoids additional
nessesary repaints and strange looking
animations.