task: introduce sftp
This patch introduces Flysystem (https://flysystem.thephpleague.com/) as
the underlying package to handle remote storage. For now, we are only
integrating SFTP in addition to FTP to allow more connection options for
FTP uploads. There will now be an option to select FTP or SFTP, and the
correct adapter will be used to create the connection.

The code has been restructured to move logic away from the endpoints
towards a central service. This will make it easier to implement and
maintain in the future.

Some options have been removed as they do not seem to work reliably:

- `ftp:appendDate` is a nice idea, but it has a flawed implementation
  as the date is evaluated on each request. This results in a mismatch
  between requested and actual URLs to files.

- `ftp:upload_thumb`: Thumbnails are now always uploaded, ensuring all
  files are available on the server for the generated webpage.
  Defaults were introduced for several config settings. This way, we can
  rely on the settings being set and do not need to guess.

The template for the server has been restructured. Instead of parsing
the template and potentially risking breaking it, we are now providing
a dedicated config file that can be loaded without problems.
Additionally, a test template that uses local data is now available.