Ensure live preview is not cached
On some devices, the video preview freezes when the video is hidden and
then shown again (e.g, show preview during countdown; hide preview to
show taken picture; show preview during countdown of the next picture).
This is because the last frame can be cached by the browser when the
preview is hidden, and this cached frame is displayed when the preview
is shown again.
To disable caching, a unique parameter is added to the URL of the
stream of the stream.

Signed-off-by: Nils Bosbach <n.bosbach@gmx.de>